feat: migrate CreateTeamScreen to :feature:createTeam KMP module
- Create :feature:createTeam KMP module with expect/actual PlatformBackHandler
- Move CreateTeamScreenStateless, CreateTeamUiState, ValueAndValidation to commonMain
- Delete old ValueAndValidation.kt from :app
- App keeps stateful wrapper with koinViewModel
