--- app/src/main/kotlin/com/sedilant/yambol/ui/createTeam/ValueAndValidation.kt
@@ -1,6 +1,0 @@
-package com.sedilant.yambol.ui.createTeam
-
-data class ValueAndValidation (
-    val input: String = "",
-    val isValid: Boolean = true
-)
