+++ feature/createTeam/src/androidMain/kotlin/com/sedilant/yambol/feature/createTeam/PlatformBackHandler.android.kt
@@ -1,0 +1,9 @@
+package com.sedilant.yambol.feature.createTeam
+
+import androidx.activity.compose.BackHandler
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun PlatformBackHandler(enabled: Boolean, onBack: () -> Unit) {
+    BackHandler(enabled = enabled, onBack = onBack)
+}
