+++ feature/createTeam/src/commonMain/kotlin/com/sedilant/yambol/feature/createTeam/PlatformBackHandler.kt
@@ -1,0 +1,6 @@
+package com.sedilant.yambol.feature.createTeam
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun PlatformBackHandler(enabled: Boolean, onBack: () -> Unit)
