+++ feature/createTeam/src/commonMain/kotlin/com/sedilant/yambol/feature/createTeam/ValueAndValidation.kt
@@ -1,0 +1,6 @@
+package com.sedilant.yambol.feature.createTeam
+
+data class ValueAndValidation(
+    val input: String = "",
+    val isValid: Boolean = true
+)
