+++ feature/createTeam/src/iosMain/kotlin/com/sedilant/yambol/feature/createTeam/PlatformBackHandler.ios.kt
@@ -1,0 +1,8 @@
+package com.sedilant.yambol.feature.createTeam
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun PlatformBackHandler(enabled: Boolean, onBack: () -> Unit) {
+    // No-op on iOS — system back gesture not applicable
+}
